Removed unnecessary '\n' from the activity status.
This won't fix any actual issues if they exist, but will remove a visual issue that props up on mobile devices from a new line in the status. With a newline, it adds exclaimation marks around the status, which doesn't look great.  Removing the newline character at the end fixes this issue.